Doctorsmove/requirements.txt
Flask
mysql-connector-python
boto3
requests
PyMySQL
pythonping
Werkzeug
flask-bcrypt